Put VertaModel in external module
